requirements.txt
streamlit
langchain
tiktoken==0.4.0
unstructured==0.6.8
tabulate==0.9.0
pdf2image==1.16.3
pytesseract==0.3.10
streamlit
mistralai
python-dotenv
streamlit-ace
ai71
